backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib==1.7.4 
bcrypt==3.2.0
python-jose[cryptography]
twilio
slowapi
httpx